% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedDada.R
\name{makeSequenceTable}
\alias{makeSequenceTable}
\title{Build Sample x ASV Count Matrix}
\usage{
makeSequenceTable(samples, orderBy = "abundance")
}
\arguments{
\item{samples}{List of \code{dada} / \code{derep} / \code{mergePairs}
data frames, optionally named.}

\item{orderBy}{\code{"abundance"} (default) sorts columns by total count;
\code{""} keeps insertion order.}
}
\value{
Integer matrix [samples x ASV sequences].
}
\description{
Drop-in replacement for \code{dada2::makeSequenceTable}.
}
\examples{
fastq <- system.file("extdata", "sample_R1.fastq", package = "SpeedDada")
d <- derepFastq(fastq)
err <- learnErrors(fastq, nbases = 1e4)
res <- dada(d, err, omega_a = 1e-5)
makeSequenceTable(list(s1 = res))

}
